Fold essentials and arrays
